Rename `sampling` module to `peer_sampling`
